Print and Plot

ATA.print_infos — Method

print_infos(ATAmodel)

Description

Print the results of each build step of the ATA model.

Arguments

ATAmodel::AbstractModel : Required. Model processed with build functions.

ATA.print_last_info — Method

print_last_info(ATAmodel)

Description

Print info of the last build step.

Arguments

ATAmodel::AbstractModel : Required. Model processed with build functions.

ATA.print_results — Method

print_results(ATAmodel; group_by_fs = false, results_folder = "RESULTS")

Description

Print the features of the assembled tests.

Arguments

ATAmodel::AbstractModel : Required. The model built with ATA fuctions, ATAmodel.design matrix must be IxT or nfsxT if the items are grouped by friend sets.

group_by_fs : Optional. Default: false. Set to true if items have been grouped by friend sets by group_by_friends!.

results_folder : Optional. Default: "RESULTS". The folder in which the output is stored.