Print and Plot

ATA.add_constraints! — Method

addconstraints!(atamodel::AbstractModel; constraintsfile = "constraints.csv", constraintsdelim = ";")

Description

Add categorical and sum constraints to the ATA.AbstractModel as specified in the constraints_file. Alternatively, the constraint dataframe can be passed using the argument constraints. It requires the start_ata step.

Arguments

ata_model::AbstractModel : Required. The model built with start_ata() and with settings loaded by start_ata function.

`constraints::DataFrames.DataFrame = DataFrames.DataFrame()`. Optional. Default: DataFrames.DataFrame(). The dataframe containing the categorical and quantitative constraints.

constraints_file : Optional. Default: "constraints.csv". The path of the file containing the categorical and sum constraints in the form of custom-separated values.

constraints_delim : Optional. Default: ";". The custom-separator for the constraints_file.

ATA.add_enemies! — Method

addenemies!(atamodel::AbstractModel)

Description

Add enemy sets to the ATA.AbstractModel as specified in the settings_file loaded by start_ata function. It requires the start_ata step.

Arguments

ata_model::AbstractModel : Required. The model built with start_ata() and with settings loaded by start_ata function.

ATA.add_exp_score! — Method

addexpscore!(ata_model::AbstractModel)

Description

Add the expected score constraints as specified in the settings_file. If the names of the expected score columns of the item bank are not provided in the settings_file they are computed as the item characteristic functions following the IRT model always specified in the settings_file. It requires the start_ata step.

Arguments

ata_model::AbstractModel : Required. The model built with start_ata() and with settings loaded by start_ata function.

ATA.add_friends! — Method

addfriends!(atamodel::AbstractModel)

Description

Add friend sets to the ATA.AbstractModel as specified in the settings_file loaded by start_ata function. It requires the start_ata step.

Arguments

ata_model::AbstractModel : Required. The model built with start_ata() function.

ATA.add_obj_fun! — Method

add_obj_fun!(ata_model::CcMaximinModel; psychometrics = false, kwargs...)

Description

Add the objective function as specified in the settings_file. It requires the start_ata build step. Computes the IIFs at predefined ability points using R sampled item parameters.

Arguments

ata_model::CcMaximinModel : Required. The model built with start_ata() and with settings loaded by start_ata function.

ATA.add_obj_fun! — Method

add_obj_fun!(ata_model::Union{MaximinModel,MinimaxModel}; kwargs...)

Description

Add the objective function as specified in the settings_file. It requires the start_ata build step. Computes the IIFs at predefined ability points.

Arguments

ata_model::Union{MaximinModel, MinimaxModel} : Required. The model built with start_ata() and with settings loaded by start_ata function.

ATA.add_overlap! — Method

addoverlap!(atamodel::AbstractModel; overlapfile = "overlapmatrix.csv", overlap_delim=";")

Description

Add maximum overlap constraints to the ATA.AbstractModel as specified by the overlap_file which contains a TxT matrix defining the maximum number of shared items between each pair of tests. Alternatively, the overlap matrix can be passed using the argument overlap. It requires the start_ata step.

Arguments

ata_model::AbstractModel : Required. The model built with start_ata() and with settings loaded by start_ata function.

overlap::Matrix{Float64} = Matrix{Float64}(undef, 0, 0): Optional.

overlap_file : Optional. Default: "overlap_matrix.csv". The path of the file containing the maximum overlap between each pair of tests in the form of a matrix with custom-separated values.

overlap_delim : Optional. Default: ";". The custom-separator for the overlap_file.

ATA.fs_to_items — Method

fs_to_items(xₜ, n_items, fs_items)

Description

Ungroup items grouped by friend sets.

Arguments

xₜ::Vector{Float64} : Required. grouped items 0-1 vector.

n_items::Int64 : Required. Number if items.

fs_items::Vector{Vector{Int64}} : Required. vector of items included in the friend sets.

Output

x_Iₜ::Vector{Float64} Returns a 0-1 vector at item level.

ATA.group_by_friends! — Method

groupbyfriends!(ata_model::AbstractModel)

Description

Group the items by friend sets once the friend sets have been added to the ATA.AbstractModel by add_friends! function and as last operation of the assembly.

Arguments

ata_model::AbstractModel : Required. The model built with start_ata(), settings loaded by start_ata function and friend sets added by add_friends! function.

ATA.item_char — Method

item_char(pars::DataFrames.DataFrame, theta::Float64; model = "2PL", parametrization = "at-b", D = 1, derivatives = false)

Description

Compute the item characteristic function (probability of correct answer).

Arguments

pars::DataFrames.DataFrame : Required. DataFrame with item parameters (difficulty: b or d, discrimination: a or a1, guessing: c or g).

theta::Float64 : Required. Ability point.

model : Optional. Default: "2PL". Values: "1PL", "2PL", "3PL". IRT model.

parametrization : Optional. Default: "at-b". Values: "at-b", "at-ab", "at+b", "at+ab". IRT model parametrization. Ex: at-b is $Da(\theta-b)$.

D : Optional. Default: 1.

derivatives : Optional. Default: false. If it's true compute the derivatives. Ow a zeros(0,0) matrix is returned.

Output

p::Matrix{Float64}: Matrix of probabilites.

pder::Matrix{Float64}: Matrix of derivatives.

ATA.item_char — Method

item_char(pars::DataFrames.DataFrame, theta::Vector{Float64}; model = "2PL", parametrization = "at-b", D = 1, derivatives = false)

Description

Compute the item characteristic function (probability of correct answer).

Arguments

pars::DataFrames.DataFrame : Required. DataFrame with item parameters (difficulty: b or d, discrimination: a or a1, guessing: c or g).

theta::Vector{Float64} : Required. Vector of ability points.

model : Optional. Default: "2PL". Values: "1PL", "2PL", "3PL". IRT model.

parametrization : Optional. Default: "at-b". Values: "at-b", "at-ab", "at+b", "at+ab". IRT model parametrization. Ex: at-b is $Da(\theta-b)$.

D : Optional. Default: 1.

derivatives : Optional. Default: false. If it's true compute the derivatives. Ow a zeros(0,0) matrix is returned.

Output

p::Matrix{Float64}: Matrix of probabilites.

pder::Matrix{Float64}: Matrix of derivatives.

ATA.item_info — Method

item_info(
    pars::DataFrames.DataFrame,
    theta::Float64;
    model = "2PL", #1PL, 2PL, 3PL, grm
    parametrization = "at-b", #"at-b, at-ab, at+b, at+ab"
    D = 1,
)

Description

Compute the item Fisher information function.

Arguments

pars::DataFrames.DataFrame : Required. DataFrame with item parameters (difficulty: b or d, discrimination: a or a1, guessing: c or g).

theta::Float64 : Required. Vector of ability points.

model : Optional. Default: "2PL". Values: "1PL", "2PL", "3PL". IRT model.

parametrization : Optional. Default: "at-b". Values: "at-b", "at-ab", "at+b", "at+ab". IRT model parametrization. Ex: at-b is $Da(\theta-b)$.

D : Optional. Default: 1.

Output

i::Matrix{Float64}: Matrix of the item information.

ATA.item_info — Method

item_info(
    pars::DataFrames.DataFrame,
    theta::Vector{Float64};
    model = "2PL", #1PL, 2PL, 3PL, grm
    parametrization = "at-b", #"at-b, at-ab, at+b, at+ab"
    D = 1,
)

Description

Compute the item Fisher information function.

Arguments

pars::DataFrames.DataFrame : Required. DataFrame with item parameters (difficulty: b or d, discrimination: a or a1, guessing: c or g).

theta::Vector{Float64} : Required. Vector of ability points.

model : Optional. Default: "2PL". Values: "1PL", "2PL", "3PL". IRT model.

parametrization : Optional. Default: "at-b". Values: "at-b", "at-ab", "at+b", "at+ab". IRT model parametrization. Ex: at-b is $Da(\theta-b)$.

D : Optional. Default: 1.

Output

i::Matrix{Float64}: Matrix of the item information.

ATA.print_infos — Method

printinfos(atamodel)

Description

Print the results of each build step of the ATA model.

Arguments

ata_model::AbstractModel : Required. Model processed with build functions.

ATA.print_last_info — Method

printlastinfo(ata_model)

Description

Print info of the last build step.

Arguments

ata_model::AbstractModel : Required. Model processed with build functions.

ATA.print_results — Method

print_results(
    ata_model::CcMaximinModel;
    group_by_fs = false,
    results_folder = "results",
    sim_pool::DataFrames.DataFrame = DataFrame(),
)

Description

Print the features of the assembled tests.

Arguments

ata_model::CcMaximinModel : Required. The model built with ATA fuctions, ata_model.design matrix must be IxT or nfsxT if the items are grouped by friend sets.

group_by_fs : Optional. Default: false. Set to true if items have been grouped by friend sets by group_by_friends!.

results_folder : Optional. Default: "results". The folder in which the output is stored.

sim_pool::DataFrames.DataFrame : Optional. Default: DataFrame(). The pool with true item paramaters. For simulation studies.

ATA.print_results — Method

print_results(
    ata_model::MaximinModel;
    group_by_fs = false,
    results_folder = "results",
    sim_pool::DataFrames.DataFrame = DataFrame(),
)

Description

Print the features of the assembled tests.

Arguments

ata_model::MaximinModel : Required. The model built with ATA fuctions, ata_model.design matrix must be IxT or nfsxT if the items are grouped by friend sets.

group_by_fs : Optional. Default: false. Set to true if items have been grouped by friend sets by group_by_friends!.

results_folder : Optional. Default: "results". The folder in which the output is stored.

sim_pool::DataFrames.DataFrame : Optional. Default: DataFrame(). The pool with true item paramaters. For simulation studies.

ATA.print_results — Method

print_results(
    ata_model::NoObjModel;
    group_by_fs = false,
    results_folder = "results",
    sim_pool::DataFrames.DataFrame = DataFrame(),
)

Description

Print the features of the assembled tests.

Arguments

ata_model::NoObjModel : Required. The model built with ATA fuctions, ata_model.design matrix must be IxT or nfsxT if the items are grouped by friend sets.

group_by_fs : Optional. Default: false. Set to true if items have been grouped by friend sets by group_by_friends!.

results_folder : Optional. Default: "results". The folder in which the output is stored.

sim_pool::DataFrames.DataFrame : Optional. Default: DataFrame(). The pool with true item paramaters. For simulation studies.

ATA.resp_gen — Method

resp_gen( pars::DataFrames.DataFrame, theta::Vector{Float64}, design::Matrix{Float64}; model = "2PL", #1PL, 2PL, 3PL parametrization = "at-b", #"at-b, at-ab, at+b, at+ab" D = 1,#any number method = "classic uniform", )

Description

Generate dichotomous responses from a dataframe of item parameters and a vector of ability points (theta).

Arguments

pars::DataFrames.DataFrame : Required. DataFrame with item parameters (difficulty: b or d, discrimination: a or a1, guessing: c or g).

theta::Vector{Float64} : Required. Vector of ability points.

design::Matrix{Float64} : Required. The design matrix.

model: Optional. Default

parametrization : Optional. Default: "at-b". Values: "at-b", "at-ab", "at+b", "at+ab". IRT model parametrization. Ex: at-b is $Da(\theta-b)$.

D: Optional. Default: 1.

method: Optional. Default: "classic uniform".

Output

i::Matrix{Float64}: Matrix of binary responses.

ATA.start_ata — Method

start_ata(;
    settings::InputSettings = InputSettings(),
    bank::DataFrames.DataFrame = DataFrames.DataFrame(),
    settings_file = "settings_ata.jl",
    bank_file = "bank.csv",
    bank_delim = ";",
)

Description

Initialize an empty ATA model and load the test assembly settings contained in the settings_file using item data in file bank_file which values are separated by bank_delim. Alternatively, settings object and bank dataframe can be passed using the arguments settings and bank.

Arguments

settings::InputSettings : Optional. Default: InputSettings() object with custom ATA settings.

bank::DataFrames.DataFrame : Optional. Default: DataFrame(). A dataframe containing data about the items.

settings_file : Optional. Default: "settings_ata.jl". The path of the file containing the ATA settings in the form of an InputSettings struct.

bank_file : Optional. Default: "bank.csv". The path of the file containing the item pool/bank in the form of custom-separated values.

bank_delim : Optional. Default: ";". The custom-separator for the bank_file.

Output

An ATA model.