Build the ATA model

ATA.start_ATA — Method

start_ATA()

Description

Start the test assembly instance.

Output

It returns an empty ATA.Model object.

ATA.load_settings! — Method

load_settings!(ATAmodel::Model; settings_file = "SettingsATA.jl", bank_file = "Bank.csv", bank_delim = ";")

Description

Load the test assembly settings contained in the settings_file. It requires the start_ATA step.

Arguments

ATAmodel::Model : Required. The model built with start_ATA() function.

settings_file : Optional. Default: "SettingsATA.jl". The path of the file containing the ATA settings in the form of an InputSettings struct.

bank_file : Optional. Default: "Bank.csv". The path of the file containing the item pool/bank in the form of custom-separated values.

bank_delim : Optional. Default: ";". The custom-separator for the bank_file.

ATA.add_friends! — Method

add_friends!(ATAmodel::Model)

Description

Add friend sets to the ATA.Model as specified in the settings_file loaded by load_settings! function. It requires the load_settings! step.

Arguments

ATAmodel::Model : Required. The model built with start_ATA() and with settings loaded by load_settings! function.

ATA.add_enemies! — Method

add_enemies!(ATAmodel::Model)

Description

Add enemy sets to the ATA.Model as specified in the settings_file loaded by load_settings! function. It requires the load_settings! step.

Arguments

ATAmodel::Model : Required. The model built with start_ATA() and with settings loaded by load_settings! function.

ATA.add_constraints! — Method

add_constraints!(ATAmodel::Model; constraints_file = "Constraints.csv", constraints_delim = ";")

Description

Add categorical and sum constraints to the ATA.Model as specified in the constraints_file. It requires the load_settings! step.

Arguments

ATAmodel::Model : Required. The model built with start_ATA() and with settings loaded by load_settings! function.

constraints_file : Optional. Default: "Constraints.csv". The path of the file containing the categorical and sum constraints in the form of custom-separated values.

constraints_delim : Optional. Default: ";". The custom-separator for the constraints_file.

ATA.add_overlap! — Method

add_overlap!(ATAmodel::Model; overlap_file = "OverlapMatrix.csv", overlap_delim=";")

Description

Add maximum overlap constraints to the ATA.Model as specified by the overlap_file which contains a TxT matrix defining the maximum number of shared items between each pair of tests. It requires the load_settings! step.

Arguments

ATAmodel::Model : Required. The model built with start_ATA() and with settings loaded by load_settings! function.

overlap_file : Optional. Default: "OverlapMatrix.csv". The path of the file containing the maximum overlap between each pair of tests in the form of a matrix with custom-separated values.

overlap_delim : Optional. Default: ";". The custom-separator for the overlap_file.

ATA.add_exp_score! — Method

add_exp_score!(ATAmodel::Model)

Description

Add the expected score constraints as specified in the settings_file. If the names of the expected score columns of the item bank are not provided in the settings_file they are computed as the item characteristic functions following the IRT model always specified in the settings_file. It requires the load_settings! step.

Arguments

ATAmodel::Model : Required. The model built with start_ATA() and with settings loaded by load_settings! function.

ATA.add_obj_fun! — Method

add_obj_fun!(ATAmodel::Model)

Description

Add the objective function as specified in the settings_file. It requires the load_settings! step.

Arguments

ATAmodel::Model : Required. The model built with start_ATA() and with settings loaded by load_settings! function.

ATA.group_by_friends! — Method

group_by_friends!(ATAmodel::Model)

Description

Group the items by friend sets once the friend sets have been added to the ATA.Model by add_friends! function and as last operation of the assembly.

Arguments

ATAmodel::Model : Required. The model built with start_ATA(), settings loaded by load_settings! function and friend sets added by add_friends! function.

ATA.load_design! — Method

load_design!(design::Matrix{Any}, ATAmodel::Model)

Description

Load a 0-1 IxT (or nFSxT if the items are grouped by friend sets) design matrix into the ATA model. Useful for loading a custom starting design before the assemble! step or to print/plot the features of the tests produced by a custom design before running print_results

Arguments

ATAmodel::Model : Required. The model built with start_ATA(), settings loaded by load_settings!.