Print and Plot

ATA.print_results — Method

print_results(ATAmodel; group_by_fs = false, results_folder = "RESULTS")

Description

Print the features of the assembled tests.

Arguments

ATAmodel::Model : Required. The model built with ATA fuctions, ATAmodel.design matrix must be IxT or nFSxT if the items are grouped by friend sets.

group_by_fs : Optional. Default: false. Set to true if items have been grouped by friend sets by group_by_friends!.

results_folder : Optional. Default: "RESULTS". The folder in which the output is stored.