Print and Plot

Print Results on Text File

ATA.print_results — Method

print_results(
    ata_model::NoObjModel;
    group_by_fs = false,
    results_folder = "results",
    sim_pool::DataFrames.DataFrame = DataFrame(),
)

Description

Print the features of the assembled tests.

Arguments

ata_model::NoObjModel : Required. The model built with ATA fuctions, ata_model.design matrix must be IxT or nfsxT if the items are grouped by friend sets.

group_by_fs : Optional. Default: false. Set to true if items have been grouped by friend sets by group_by_friends!.

results_folder : Optional. Default: "results". The folder in which the output is stored.

sim_pool::DataFrames.DataFrame : Optional. Default: DataFrame(). The pool with true item paramaters. For simulation studies.

ATA.print_results — Method

print_results(
    ata_model::Union{MaximinModel, SoysterMaximinModel, DeJongMaximinModel};
    group_by_fs = false,
    results_folder = "results",
    sim_pool::DataFrames.DataFrame = DataFrame(),
)

Description

Print the features of the assembled tests.

Arguments

ata_model::Union{MaximinModel, SoysterMaximinModel, DeJongMaximinModel} : Required. The model built with ATA fuctions, ata_model.design matrix must be IxT or nfsxT if the items are grouped by friend sets.

group_by_fs : Optional. Default: false. Set to true if items have been grouped by friend sets by group_by_friends!.

results_folder : Optional. Default: "results". The folder in which the output is stored.

sim_pool::DataFrames.DataFrame : Optional. Default: DataFrame(). The pool with true item paramaters. For simulation studies.

Missing docstring.

Missing docstring for print_results( ata_model::MinimaxModel; group_by_fs = false, results_folder = "results", sim_pool::DataFrames.DataFrame = DataFrame(), ). Check Documenter's build log for details.

ATA.print_results — Method

print_results(
    ata_model::CCMaximinModel;
    group_by_fs = false,
    results_folder = "results",
    sim_pool::DataFrames.DataFrame = DataFrame(),
)

Description

Print the features of the assembled tests.

Arguments

ata_model::CCMaximinModel : Required. The model built with ATA fuctions, ata_model.design matrix must be IxT or nfsxT if the items are grouped by friend sets.

group_by_fs : Optional. Default: false. Set to true if items have been grouped by friend sets by group_by_friends!.

results_folder : Optional. Default: "results". The folder in which the output is stored.

sim_pool::DataFrames.DataFrame : Optional. Default: DataFrame(). The pool with true item paramaters. For simulation studies.

Plot TIFs and ICFs

Missing docstring.

Missing docstring for plot_results( ata_model::AbstractModel; group_by_fs = false, results_folder = "plots", ). Check Documenter's build log for details.